Extract `backup-node-modules.js`
