+++ script/test
@@ -11,6 +11,7 @@ const glob = require('glob')
 const path = require('path')
 
 const CONFIG = require('./config')
+const backupNodeModules = require('./lib/backup-node-modules')
 const runApmInstall = require('./lib/run-apm-install')
 
 const resourcePath = CONFIG.repositoryRootPath
@@ -94,11 +95,7 @@ for (let packageName in CONFIG.appMetadata.packageDependencies) {
     if (require(pkgJsonPath).atomTestRunner) {
       console.log(`Installing test runner dependencies for ${packageName}`.bold.green)
       if (fs.existsSync(nodeModulesPath)) {
-        fs.copySync(nodeModulesPath, nodeModulesBackupPath)
-        finalize = () => {
-          fs.removeSync(nodeModulesPath)
-          fs.renameSync(nodeModulesBackupPath, nodeModulesPath)
-        }
+        finalize = backupNodeModules(repositoryPackagePath)
       } else {
         finalize = () => fs.removeSync(nodeModulesPath)
       }
